build: add plugin
add a build solution to the project that buids a plugin zip that can be
published on the marketplace. Most values in the plugin descriptor are
their default values and would need be changed when publishing to the
marketplace.

To run the build use the new runconfiguration that is part of this
commit. The generated build.xml could be used in the future for something
like github actions to automate the build. Though that would require
some more work like downloading MPS for the build etc.